+++ express_relay/sdk/js/src/types.ts
@@ -101,29 +101,6 @@ export type Opportunity = {
  */
 export type OpportunityParams = Omit<Opportunity, "opportunityId">;
 /**
- * Represents a bid for an opportunity
- */
-export type OpportunityBid = {
-  /**
-   * Opportunity unique identifier in uuid format
-   */
-  opportunityId: string;
-  /**
-   * The permission key required for successful execution of the opportunity.
-   */
-  permissionKey: Hex;
-  /**
-   * Executor address
-   */
-  executor: Address;
-  /**
-   * Signature of the executor
-   */
-  signature: Hex;
-
-  bid: BidParams;
-};
-/**
  * Represents a raw bid on acquiring a permission key
  */
 export type Bid = {
@@ -161,108 +138,4 @@ export type BidStatusUpdate = {
 export type BidResponse = components["schemas"]["SimulatedBid"];
 export type BidsResponse = {
   items: BidResponse[];
-};
-
-export const executeOpportunityAbi = {
-  type: "function",
-  name: "executeOpportunity",
-  inputs: [
-    {
-      name: "params",
-      type: "tuple",
-      internalType: "struct ExecutionParams",
-      components: [
-        {
-          name: "permit",
-          type: "tuple",
-          internalType: "struct ISignatureTransfer.PermitBatchTransferFrom",
-          components: [
-            {
-              name: "permitted",
-              type: "tuple[]",
-              internalType: "struct ISignatureTransfer.TokenPermissions[]",
-              components: [
-                {
-                  name: "token",
-                  type: "address",
-                  internalType: "address",
-                },
-                {
-                  name: "amount",
-                  type: "uint256",
-                  internalType: "uint256",
-                },
-              ],
-            },
-            {
-              name: "nonce",
-              type: "uint256",
-              internalType: "uint256",
-            },
-            {
-              name: "deadline",
-              type: "uint256",
-              internalType: "uint256",
-            },
-          ],
-        },
-        {
-          name: "witness",
-          type: "tuple",
-          internalType: "struct ExecutionWitness",
-          components: [
-            {
-              name: "buyTokens",
-              type: "tuple[]",
-              internalType: "struct TokenAmount[]",
-              components: [
-                {
-                  name: "token",
-                  type: "address",
-                  internalType: "address",
-                },
-                {
-                  name: "amount",
-                  type: "uint256",
-                  internalType: "uint256",
-                },
-              ],
-            },
-            {
-              name: "executor",
-              type: "address",
-              internalType: "address",
-            },
-            {
-              name: "targetContract",
-              type: "address",
-              internalType: "address",
-            },
-            {
-              name: "targetCalldata",
-              type: "bytes",
-              internalType: "bytes",
-            },
-            {
-              name: "targetCallValue",
-              type: "uint256",
-              internalType: "uint256",
-            },
-            {
-              name: "bidAmount",
-              type: "uint256",
-              internalType: "uint256",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      name: "signature",
-      type: "bytes",
-      internalType: "bytes",
-    },
-  ],
-  outputs: [],
-  stateMutability: "payable",
 };
